Gathered common types to one file
